行動力変更処理を追加した

+++ Assets/MyGame/Scripts/Piece/PieceParameter.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public class PieceParameter : MonoBehaviour
 {
-    public float Activity { get; private set; }
+    public int Activity { get; private set; }
 }
